Implemented Client side of pangolin Cli
Used pangolin-core core to build the client side of pangolin.
Forced the creation of a Pangolin Frame for manual verification.
Fixed argo-uml exception. The log4j initialization raised a Null Pointer
Exception, because the System properties value was not in agreement with
the actual system. So we inject code to initialize the value as Null.

Signed-off-by: Bruno Sotto-Mayor